Hearo/source_separation/pixel_ring/requirements.txt
spidev
pyusb==1.1.0
